Integrated frontend and backend. Existing backend code moved to Backend.js, frontend code soon-to-be integrated into Home.js. We should use different documents for the Home and the Contacts page
