+++ src/Home.js
@@ -1,0 +1,9 @@
+import "./styles.css"
+
+function Home() {
+	return(
+		<div className="Test">This is a test of the Home.js component. </div>
+	)
+}
+
+export default Home;
