+++ src/styles.css
@@ -1,0 +1,5 @@
+div.Test{
+	color: red;
+	font-family: "serif";
+
+}
